Simplify compiler implementation to not depend on vm
