+++ lib/cli/Cargo.toml
@@ -34,7 +34,6 @@ wasmer-engine = { version = "2.0.0", path = "../engine" }
 wasmer-engine-universal = { version = "2.0.0", path = "../engine-universal", optional = true }
 wasmer-engine-dylib = { version = "2.0.0", path = "../engine-dylib", optional = true }
 wasmer-engine-staticlib = { version = "2.0.0", path = "../engine-staticlib", optional = true }
-wasmer-vm = { version = "2.0.0", path = "../vm" }
 wasmer-wasi = { version = "2.0.0", path = "../wasi", optional = true }
 wasmer-wasi-experimental-io-devices = { version = "2.0.0", path = "../wasi-experimental-io-devices", optional = true }
 wasmer-wast = { version = "2.0.0", path = "../../tests/lib/wast", optional = true }
