+++ lib/compiler-cranelift/Cargo.toml
@@ -13,7 +13,6 @@ edition = "2018"
 
 [dependencies]
 wasmer-compiler = { path = "../compiler", version = "2.0.0", features = ["translator"], default-features = false }
-wasmer-vm = { path = "../vm", version = "2.0.0" }
 wasmer-types = { path = "../types", version = "2.0.0", default-features = false, features = ["std"] }
 cranelift-entity = { version = "0.76", default-features = false }
 cranelift-codegen = { version = "0.76", default-features = false, features = ["x86", "arm64"] }
