+++ lib/compiler-cranelift/src/compiler.rs
@@ -272,7 +272,7 @@ impl Compiler for CraneliftCompiler {
             .into_iter()
             .collect::<PrimaryMap<SignatureIndex, FunctionBody>>();
 
-        use wasmer_vm::VMOffsets;
+        use wasmer_types::VMOffsets;
         let offsets = VMOffsets::new_for_trampolines(frontend_config.pointer_bytes());
         // dynamic function trampolines (only for imported functions)
         let dynamic_function_trampolines = module
