+++ lib/compiler-cranelift/src/trampoline/dynamic_function.rs
@@ -17,8 +17,7 @@ use std::mem;
 
 use cranelift_frontend::{FunctionBuilder, FunctionBuilderContext};
 use wasmer_compiler::{CompileError, FunctionBody};
-use wasmer_types::FunctionType;
-use wasmer_vm::VMOffsets;
+use wasmer_types::{FunctionType, VMOffsets};
 
 /// Create a trampoline for invoking a WebAssembly function.
 pub fn make_trampoline_dynamic_function(
