+++ lib/compiler-llvm/Cargo.toml
@@ -13,7 +13,6 @@ edition = "2018"
 
 [dependencies]
 wasmer-compiler = { path = "../compiler", version = "2.0.0", features = ["translator"] }
-wasmer-vm = { path = "../vm", version = "2.0.0" }
 wasmer-types = { path = "../types", version = "2.0.0" }
 target-lexicon = { version = "0.12.2", default-features = false }
 smallvec = "1.6"
