+++ lib/compiler-llvm/src/translator/intrinsics.rs
@@ -26,9 +26,9 @@ use wasmer_compiler::CompileError;
 use wasmer_types::entity::{EntityRef, PrimaryMap};
 use wasmer_types::{
     FunctionIndex, FunctionType as FuncType, GlobalIndex, LocalFunctionIndex, MemoryIndex,
-    ModuleInfo as WasmerCompilerModule, Mutability, SignatureIndex, TableIndex, Type,
+    MemoryStyle, ModuleInfo as WasmerCompilerModule, Mutability, SignatureIndex, TableIndex,
+    TrapCode, Type, VMBuiltinFunctionIndex, VMOffsets,
 };
-use wasmer_vm::{MemoryStyle, TrapCode, VMBuiltinFunctionIndex, VMOffsets};
 
 pub fn type_to_llvm_ptr<'ctx>(
     intrinsics: &Intrinsics<'ctx>,
