+++ lib/compiler/src/module.rs
@@ -5,7 +5,7 @@ use rkyv::{Archive, Deserialize as RkyvDeserialize, Serialize as RkyvSerialize};
 #[cfg(feature = "enable-serde")]
 use serde::{Deserialize, Serialize};
 use wasmer_types::entity::PrimaryMap;
-use wasmer_types::{Features, MemoryIndex, ModuleInfo, TableIndex, MemoryStyle, TableStyle};
+use wasmer_types::{Features, MemoryIndex, MemoryStyle, ModuleInfo, TableIndex, TableStyle};
 
 /// The required info for compiling a module.
 ///
