+++ lib/engine-dylib/Cargo.toml
@@ -13,7 +13,7 @@ edition = "2018"
 [dependencies]
 wasmer-types = { path = "../types", version = "2.0.0" }
 wasmer-compiler = { path = "../compiler", version = "2.0.0" }
-wasmer-vm = { path = "../vm", version = "2.0.0", features = ["enable-rkyv"] }
+wasmer-vm = { path = "../vm", version = "2.0.0" }
 wasmer-engine = { path = "../engine", version = "2.0.0" }
 wasmer-object = { path = "../object", version = "2.0.0" }
 serde = { version = "1.0", features = ["derive", "rc"] }
