+++ lib/engine-dylib/src/artifact.rs
@@ -37,12 +37,11 @@ use wasmer_types::entity::{BoxedSlice, PrimaryMap};
 #[cfg(feature = "compiler")]
 use wasmer_types::DataInitializer;
 use wasmer_types::{
-    FunctionIndex, LocalFunctionIndex, MemoryIndex, ModuleInfo, OwnedDataInitializer,
-    SignatureIndex, TableIndex, MemoryStyle, TableStyle, 
+    FunctionIndex, LocalFunctionIndex, MemoryIndex, MemoryStyle, ModuleInfo, OwnedDataInitializer,
+    SignatureIndex, TableIndex, TableStyle,
 };
 use wasmer_vm::{
-    FuncDataRegistry, FunctionBodyPtr, VMFunctionBody,
-    VMSharedSignatureIndex, VMTrampoline,
+    FuncDataRegistry, FunctionBodyPtr, VMFunctionBody, VMSharedSignatureIndex, VMTrampoline,
 };
 
 /// A compiled Wasm module, ready to be instantiated.
