+++ lib/engine-universal/Cargo.toml
@@ -13,7 +13,7 @@ edition = "2018"
 [dependencies]
 wasmer-types = { path = "../types", version = "2.0.0", features = ["enable-rkyv"] }
 wasmer-compiler = { path = "../compiler", version = "2.0.0", features = ["translator", "enable-rkyv"] }
-wasmer-vm = { path = "../vm", version = "2.0.0", features = ["enable-rkyv"] }
+wasmer-vm = { path = "../vm", version = "2.0.0" }
 wasmer-engine = { path = "../engine", version = "2.0.0" }
 # flexbuffers = { path = "../../../flatbuffers/rust/flexbuffers", version = "0.1.0" }
 region = "3.0"
