+++ lib/engine/src/resolver.rs
@@ -4,12 +4,14 @@
 use crate::{Export, ExportFunctionMetadata, ImportError, LinkError};
 use more_asserts::assert_ge;
 use wasmer_types::entity::{BoxedSlice, EntityRef, PrimaryMap};
-use wasmer_types::{ExternType, FunctionIndex, ImportIndex, MemoryIndex, ModuleInfo, TableIndex, MemoryStyle, TableStyle};
+use wasmer_types::{
+    ExternType, FunctionIndex, ImportIndex, MemoryIndex, MemoryStyle, ModuleInfo, TableIndex,
+    TableStyle,
+};
 
 use wasmer_vm::{
-    FunctionBodyPtr, ImportFunctionEnv, Imports, VMFunctionBody,
-    VMFunctionEnvironment, VMFunctionImport, VMFunctionKind, VMGlobalImport, VMMemoryImport,
-    VMTableImport,
+    FunctionBodyPtr, ImportFunctionEnv, Imports, VMFunctionBody, VMFunctionEnvironment,
+    VMFunctionImport, VMFunctionKind, VMGlobalImport, VMMemoryImport, VMTableImport,
 };
 
 /// Import resolver connects imports with available exported values.
