+++ lib/types/Cargo.toml
@@ -16,6 +16,7 @@ thiserror = "1.0"
 indexmap = { version = "1.6", features = ["serde-1"] }
 rkyv = { version = "0.6.1", optional = true }
 loupe = { version = "0.1", features = ["enable-indexmap"] }
+more-asserts = "0.2"
 
 [features]
 default = ["std", "enable-serde", "enable-rkyv"]
