+++ lib/types/src/vm/mod.rs
@@ -1,0 +1,5 @@
+mod styles;
+mod vmoffsets;
+
+pub use styles::{MemoryStyle, TableStyle};
+pub use vmoffsets::{VMBuiltinFunctionIndex, VMOffsets};
