+++ lib/vm/src/export.rs
@@ -8,7 +8,7 @@ use crate::table::Table;
 use crate::vmcontext::{VMFunctionBody, VMFunctionEnvironment, VMFunctionKind, VMTrampoline};
 use loupe::MemoryUsage;
 use std::sync::Arc;
-use wasmer_types::{FunctionType, MemoryType, MemoryStyle, TableType, TableStyle};
+use wasmer_types::{FunctionType, MemoryStyle, MemoryType, TableStyle, TableType};
 
 /// The value of an export passed from one instance to another.
 #[derive(Debug)]
