+++ lib/vm/src/instance/allocator.rs
@@ -1,12 +1,11 @@
 use super::{Instance, InstanceRef};
 use crate::vmcontext::{VMMemoryDefinition, VMTableDefinition};
-use crate::VMOffsets;
 use std::alloc::{self, Layout};
 use std::convert::TryFrom;
 use std::mem;
 use std::ptr::{self, NonNull};
 use wasmer_types::entity::EntityRef;
-use wasmer_types::{LocalMemoryIndex, LocalTableIndex, ModuleInfo};
+use wasmer_types::{LocalMemoryIndex, LocalTableIndex, ModuleInfo, VMOffsets};
 
 /// This is an intermediate type that manages the raw allocation and
 /// metadata when creating an [`Instance`].
