+++ lib/vm/src/instance/mod.rs
@@ -26,7 +26,7 @@ use crate::vmcontext::{
     VMMemoryDefinition, VMMemoryImport, VMSharedSignatureIndex, VMTableDefinition, VMTableImport,
     VMTrampoline,
 };
-use crate::{FunctionBodyPtr, VMOffsets};
+use crate::FunctionBodyPtr;
 use crate::{VMFunction, VMGlobal, VMMemory, VMTable};
 use loupe::{MemoryUsage, MemoryUsageTracker};
 use memoffset::offset_of;
@@ -45,7 +45,7 @@ use wasmer_types::entity::{packed_option::ReservedValue, BoxedSlice, EntityRef, 
 use wasmer_types::{
     DataIndex, DataInitializer, ElemIndex, ExportIndex, FunctionIndex, GlobalIndex, GlobalInit,
     LocalFunctionIndex, LocalGlobalIndex, LocalMemoryIndex, LocalTableIndex, MemoryIndex,
-    ModuleInfo, Pages, SignatureIndex, TableIndex, TableInitializer, TrapCode,
+    ModuleInfo, Pages, SignatureIndex, TableIndex, TableInitializer, TrapCode, VMOffsets,
 };
 
 /// The function pointer to call with data and an [`Instance`] pointer to
